fix: Article commentCount/viewCount 타입을 DB INT와 일치하도록 long → int 변경

+++ src/main/java/com/sprint/mission/monew/domain/article/dto/ArticleDto.java
@@ -11,6 +11,6 @@ public record ArticleDto(
     String title,
     Instant publishDate,
     String summary,
-    long commentCount,
-    long viewCount,
+    int commentCount,
+    int viewCount,
     boolean viewedByMe) {}
